requirements.txt
streamlit
openai==0.28
geopy
Pillow
streamlit-js-eval>=0.1.6